Rename sunSpecModelFactory to modbusModelFactory
