+++ src/modbus/modbusModelFactory.ts
@@ -1,5 +1,8 @@
-import { objectEntriesWithType } from '../../helpers/object.js';
-import type { ModelAddress, SunSpecConnection } from '../connection/base.js';
+import { objectEntriesWithType } from '../helpers/object.js';
+import type {
+    ModelAddress,
+    SunSpecConnection,
+} from '../sunspec/connection/base.js';
 
 export type Mapping<
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -16,7 +19,7 @@ export type Mapping<
         : { writeConverter?: undefined });
 };
 
-export function sunSpecModelFactory<
+export function modbusModelFactory<
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
     Model extends Record<string, any>,
     WriteableKeys extends keyof Model = never,
@@ -37,7 +40,9 @@ export function sunSpecModelFactory<
     return {
         read: async ({ modbusConnection, address }) => {
             const logger = modbusConnection.logger.child({
-                module: `sunspec-model-${config.name}`,
+                module: 'modbusModelFactory',
+                model: config.name,
+                type: 'read',
             });
 
             logger.trace({ address }, 'Reading model');
@@ -59,7 +64,9 @@ export function sunSpecModelFactory<
         },
         write: async ({ modbusConnection, address, values }) => {
             const logger = modbusConnection.logger.child({
-                module: `sunspec-model-${config.name}`,
+                module: 'modbusModelFactory',
+                model: config.name,
+                type: 'write',
             });
 
             logger.trace({ address, values }, 'Writing model');
@@ -160,7 +167,7 @@ export function convertWriteRegisters<
     mapping: Mapping<Model, WriteableKeys>;
     length: number;
 }): number[] {
-    // sunspec allows for writing values to registers that do not support writing, they will simply be ignored
+    // modbus allows for writing values to registers that do not support writing, they will simply be ignored
     // we use this behaviour as a shortcut to use the same model definition and start address for reading and writing
 
     // start with all empty registers
