+++ src/sunspec/models/common.ts
@@ -5,7 +5,7 @@ import {
     registersToUint16,
     registersToUint16Nullable,
 } from '../helpers/converters.js';
-import { sunSpecModelFactory } from './sunSpecModelFactory.js';
+import { modbusModelFactory } from '../../modbus/modbusModelFactory.js';
 
 // https://sunspec.org/wp-content/uploads/2021/12/SunSpec_Information_Model_Reference_20211209.xlsx
 export type CommonModel = {
@@ -27,7 +27,7 @@ export type CommonModel = {
     DA: number | null;
 };
 
-export const commonModel = sunSpecModelFactory<CommonModel>({
+export const commonModel = modbusModelFactory<CommonModel>({
     name: 'common',
     mapping: {
         ID: {
